app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
gTTS>=2.5.0
deep-translator>=1.11.0
# For --mode slrt (heavy, already installed system-wide):
# torch, tensorflow, transformers, regex, tensorboard
